Fix-Blinking-Defect
