+++ hue-console/src/main/java/com/cgm/java/console/commands/implementations/TurnOffCommand.java
@@ -10,7 +10,6 @@ import org.apache.commons.cli.CommandLine;
 import org.apache.commons.cli.HelpFormatter;
 import org.apache.commons.cli.Options;
 import org.apache.commons.collections.ListUtils;
-import org.apache.commons.lang.StringUtils;
 
 import com.cgm.java.console.commands.BridgeCommand;
 import com.cgm.java.hue.models.Light;
@@ -19,17 +18,16 @@ import com.cgm.java.hue.utilities.HueBridgeSetter;
 import com.cgm.java.utilities.lambdas.Conversion;
 
 /**
- * A command to turn on lights
+ * A command to turn off lights.
  */
-public class TurnOn extends BridgeCommand {
-    private static final String ID_OPTION = "id";
+public class TurnOffCommand extends BridgeCommand {
     private static final String NAME_OPTION = "name";
-    private static final String BRIGHTNESS_OPTION = "brightness";
+    private static final String ID_OPTION = "id";
 
     @Override
     public void usage() {
         final HelpFormatter formatter = new HelpFormatter();
-        formatter.printHelp(getName() + " <bridge IP> <hue ID>", "A command to turn on a specific light.", getOptions(), null);
+        formatter.printHelp(getName() + " <bridge IP> <hue ID>", "A command to turn off a specific light.", getOptions(), null);
     }
 
     @Override
@@ -39,13 +37,7 @@ public class TurnOn extends BridgeCommand {
         final String[] lightIds = line.getOptionValues(ID_OPTION) == null ? new String[]{} : line.getOptionValues(ID_OPTION);
         if (lightNames.length < 1 && lightIds.length < 1) {
             usage();
-            throw new IllegalArgumentException("You must specify at least one light to turn on.");
-        }
-
-        final String[] brightnesses = line.getOptionValues(BRIGHTNESS_OPTION) == null ? new String[]{} : line.getOptionValues(BRIGHTNESS_OPTION);
-        if (brightnesses.length > 1) {
-            usage();
-            throw new IllegalArgumentException("You may only specify one brightness.");
+            throw new IllegalArgumentException("You must specify at least one light to turn off.");
         }
 
         // Collect all lights and keep them in maps
@@ -53,7 +45,7 @@ public class TurnOn extends BridgeCommand {
         final Map<String, Light> nameToLightMap = lights.stream().collect(Collectors.toMap(Conversion.LIGHT_TO_NAME, (light) -> light));
         final Map<String, Light> idToLightMap = lights.stream().collect(Collectors.toMap(Conversion.LIGHT_TO_ID, (light) -> light));
         System.out.println("These lights were found: ");
-        idToLightMap.keySet().forEach(System.out::println);
+        lights.forEach(System.out::println);
         System.out.println();
 
         // Figure out which requested lights actually match the available lights, and turn them on
@@ -70,7 +62,7 @@ public class TurnOn extends BridgeCommand {
                 light = nameToLightMap.get(lightIdentifier);
             }
 
-            final State.Builder newStateBuilder = createNewState(brightnesses[0], light);
+            final State.Builder newStateBuilder = createNewState(light);
 
             hueBridgeSetter.setLightState(bridgeIp, token, light.getId().toString(), newStateBuilder.build());
         });
@@ -78,26 +70,19 @@ public class TurnOn extends BridgeCommand {
         return 0;
     }
 
-    private State.Builder createNewState(final String brightness, final Light light) {
-        final State.Builder newStateBuilder = State.newBuilder(light.getState()).setOn(true);
-        if (StringUtils.isNotEmpty(brightness)) {
-            newStateBuilder.setBri(Long.valueOf(brightness));
-        } else {
-            newStateBuilder.setBri(254L);
-        }
-        return newStateBuilder;
+    private State.Builder createNewState(final Light light) {
+        return State.newBuilder(light.getState()).setOn(false);
     }
 
     @Override
     public String getName() {
-        return "turnon";
+        return "turnoff";
     }
 
     @Override
     public Options getOptions() {
         final Options options = super.getOptions();
         options.addOption(NAME_OPTION, true, "The name of the light to turn on.");
-        options.addOption(BRIGHTNESS_OPTION, true, "The brightness to set, 0-254.");
         options.addOption(ID_OPTION, true, "The id of the light to turn on.");
         return options;
     }
