+++ hue-models/src/main/resources/avro.usesbasic/group.avsc
@@ -2,10 +2,10 @@
  "type": "record",
  "name": "Group",
  "fields": [
-     {"name": "id", "type": ["null", "string"], "default": null},
-     {"name": "name", "type": ["null", "string"], "default": null},
+     {"name": "id", "type": ["string", "null"], "default": ""},
+     {"name": "name", "type": ["string", "null"], "default": ""},
      {"name": "lights", "type": ["null", {"type": "array", "items": "string"}], "default": null},
-     {"name": "type", "type": ["null", "string"], "default": null},
+     {"name": "type", "type": ["string", "null"], "default": ""},
      {"name": "action", "type": ["null", "State"], "default": null}
  ]
 }
