+++ hue-models/src/main/resources/avro.usesbasic/sensor.avsc
@@ -2,14 +2,14 @@
  "type": "record",
  "name": "Sensor",
  "fields": [
-     {"name": "id", "type": ["null", "string"], "default": null},
-     {"name": "name", "type": ["null", "string"], "default": null},
+     {"name": "id", "type": ["string", "null"], "default": ""},
+     {"name": "name", "type": ["string", "null"], "default": ""},
      {"name": "state", "type": ["null", "SensorState"], "default": null},
      {"name": "config", "type": ["null", "Configuration"], "default": null},
-     {"name": "type", "type": ["null", "string"], "default": null},
-     {"name": "modelid", "type": ["null", "string"], "default": null},
-     {"name": "manufacturername", "type": ["null", "string"], "default": null},
-     {"name": "uniqueid", "type": ["null", "string"], "default": null},
-     {"name": "swversion", "type": ["null", "string"], "default": null}
+     {"name": "type", "type": ["string", "null"], "default": ""},
+     {"name": "modelid", "type": ["string", "null"], "default": ""},
+     {"name": "manufacturername", "type": ["string", "null"], "default": ""},
+     {"name": "uniqueid", "type": ["string", "null"], "default": ""},
+     {"name": "swversion", "type": ["string", "null"], "default": ""}
  ]
 }
