+++ hue-web/src/main/java/com/cgm/java/hue/web/GetGroupsServlet.java
@@ -1,6 +1,7 @@
 package com.cgm.java.hue.web;
 
 import java.io.IOException;
+import java.util.List;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -11,32 +12,30 @@ import javax.servlet.http.HttpServletResponse;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import com.cgm.java.hue.models.Light;
+import com.cgm.java.hue.models.Group;
+import com.cgm.java.hue.utilities.HueBridgeGetter;
 import com.cgm.java.hue.utilities.HueConfiguration;
-import com.cgm.java.hue.utilities.LightUtil;
 import com.cgm.java.hue.web.util.KnownParameterNames;
 
 /**
- * This servlet will attempt to toggle a {@link com.cgm.java.hue.models.Light}'s state. That is, it will attempt to turn
- * the light ON or OFF.
+ * This servlet will attempt to retrieve and return all {@link com.cgm.java.hue.models.Group}s currently available on
+ * the bridge
  */
 @WebServlet("/HueServlet")
-public class ToggleLight extends HttpServlet {
-    private static final Logger LOGGER = LoggerFactory.getLogger(ToggleLight.class);
+public class GetGroupsServlet extends HttpServlet {
+    private static final Logger LOGGER = LoggerFactory.getLogger(GetGroupsServlet.class);
     private static final long serialVersionUID = 1L;
+    private static final HueBridgeGetter HUE_BRIDGE_GETTER = new HueBridgeGetter();
     private static final HueConfiguration HUE_CONFIGURATION = new HueConfiguration();
-    private static final LightUtil LIGHT_UTIL = new LightUtil();
 
-    public ToggleLight() {
+    public GetGroupsServlet() {
         super();
     }
 
     protected void doGet(final HttpServletRequest request, final HttpServletResponse response) throws ServletException, IOException {
-        // Get the input light ID, toggle it
-        final String lightId = request.getParameter(KnownParameterNames.LIGHT_ID.getName());
-        LOGGER.info("Attempting to toggle the light with ID: " + lightId);
-        final Light toggledLight = LIGHT_UTIL.toggleLight(HUE_CONFIGURATION.getIP(), HUE_CONFIGURATION.getToken(), lightId);
-        LOGGER.info("State successfully toggled to ON = " + toggledLight.getState().getOn());
+        LOGGER.info("Attempting to retrieve all groups.");
+        final List<Group> groups = HUE_BRIDGE_GETTER.getGroups(HUE_CONFIGURATION.getIP(), HUE_CONFIGURATION.getToken());
+        request.setAttribute(KnownParameterNames.GROUP_LIST.getName(), groups);
     }
 
     protected void doPost(final HttpServletRequest request, final HttpServletResponse response) throws ServletException, IOException {
