+++ hue-web/src/main/java/com/cgm/java/hue/web/ToggleLightServlet.java
@@ -1,9 +1,7 @@
 package com.cgm.java.hue.web;
 
 import java.io.IOException;
-import java.util.List;
 
-import javax.servlet.RequestDispatcher;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
@@ -14,29 +12,31 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import com.cgm.java.hue.models.Light;
-import com.cgm.java.hue.utilities.HueBridgeGetter;
 import com.cgm.java.hue.utilities.HueConfiguration;
+import com.cgm.java.hue.utilities.LightUtil;
 import com.cgm.java.hue.web.util.KnownParameterNames;
 
 /**
- * This servlet will attempt to retrieve and return all {@link com.cgm.java.hue.models.Light}s currently available on
- * the bridge
+ * This servlet will attempt to toggle a {@link com.cgm.java.hue.models.Light}'s state. That is, it will attempt to turn
+ * the light ON or OFF.
  */
 @WebServlet("/HueServlet")
-public class GetLights extends HttpServlet {
-    private static final Logger LOGGER = LoggerFactory.getLogger(GetLights.class);
+public class ToggleLightServlet extends HttpServlet {
+    private static final Logger LOGGER = LoggerFactory.getLogger(ToggleLightServlet.class);
     private static final long serialVersionUID = 1L;
-    private static final HueBridgeGetter HUE_BRIDGE_GETTER = new HueBridgeGetter();
     private static final HueConfiguration HUE_CONFIGURATION = new HueConfiguration();
+    private static final LightUtil LIGHT_UTIL = new LightUtil();
 
-    public GetLights() {
+    public ToggleLightServlet() {
         super();
     }
 
     protected void doGet(final HttpServletRequest request, final HttpServletResponse response) throws ServletException, IOException {
-        LOGGER.info("Attempting to retrieve all lights.");
-        final List<Light> list = HUE_BRIDGE_GETTER.getLights(HUE_CONFIGURATION.getIP(), HUE_CONFIGURATION.getToken());
-        request.setAttribute(KnownParameterNames.LIGHT_LIST.getName(), list);
+        // Get the input light ID, toggle it
+        final String lightId = request.getParameter(KnownParameterNames.LIGHT_ID.getName());
+        LOGGER.info("Attempting to toggle the light with ID: " + lightId);
+        final Light toggledLight = LIGHT_UTIL.toggleLight(HUE_CONFIGURATION.getIP(), HUE_CONFIGURATION.getToken(), lightId);
+        LOGGER.info("State successfully toggled to ON = " + toggledLight.getState().getOn());
     }
 
     protected void doPost(final HttpServletRequest request, final HttpServletResponse response) throws ServletException, IOException {
